控件优化
1、增添键盘动作-KEY_REPEAT，并添加GLFW与引擎InputSystem的映射关系
2、优化EditText控件
-长按退格可一直删除字符（之前由于不接受REPEAT消息而无法实现）
-空串，控件高度保留原高度（之前文字高度将置0导致控件高度骤降）
3、ClipBar控件
-动画片段的开始帧与结束帧以及长度应该是int而不是float
-添加增加、减少单位帧按钮，避免片段过长，拖拽灵敏度过低无法调节

+++ Dragon/FBXMeshLoader.cpp
@@ -591,7 +591,7 @@ FbxAMatrix FBXModelLoader::GetGeometryTransformation(FbxNode* inNode)
 
 
 
-mat4 BoneAnimation::GetTransform(int key)
+mat4 BoneAnimation::GetTransform(float key)
 {
 	//从帧集合找到指定帧相邻2个帧进行线性插值
 	vec3 trans = GetKeyValue(key, m_translation);
@@ -612,7 +612,7 @@ mat4 BoneAnimation::GetTransform(int key)
 	return transform;
 }
 
-vec3 BoneAnimation::GetKeyValue(int key, vector<Key>* Keyvector[3])
+vec3 BoneAnimation::GetKeyValue(float key, vector<Key>* Keyvector[3])
 {
 	vec3 keyValue;
 	int pre = 0, next = 0;
@@ -630,7 +630,7 @@ vec3 BoneAnimation::GetKeyValue(int key, vector<Key>* Keyvector[3])
 		}
 		if (j == Keyvector[i]->size())
 			pre = next = j - 1;
-		int ratio = pre == next ? 0 : (double)(key - (*Keyvector[i])[pre].m_key) / ((*Keyvector[i])[next].m_key - (*Keyvector[i])[pre].m_key);
+		double ratio = pre == next ? 0 : (double)(key - (*Keyvector[i])[pre].m_key) / ((*Keyvector[i])[next].m_key - (*Keyvector[i])[pre].m_key);
 		keyValue[i] = (*Keyvector[i])[pre].m_value * (1 - ratio) + (*Keyvector[i])[next].m_value * ratio;
 	}
 	return keyValue;
